+++ guillotina/auth/validators.py
@@ -1,11 +1,19 @@
-from guillotina._settings import app_settings
-from guillotina.auth import find_user
-from guillotina.utils import strings_differ
-
 import hashlib
-import jwt
+import logging
 import uuid
 
+import jwt
+from guillotina import configure
+from guillotina._settings import app_settings
+from guillotina.auth import find_user
+from guillotina.component import get_utility
+from guillotina.component import query_utility
+from guillotina.interfaces import IPasswordChecker
+from guillotina.interfaces import IPasswordHasher
+from guillotina.utils import strings_differ
+
+
+logger = logging.getLogger('guillotina')
 
 class BaseValidator(object):
     for_validators = None
@@ -14,15 +22,13 @@ class BaseValidator(object):
         self.request = request
 
 
-hash_functions: dict = {
-    'sha512': hashlib.sha512
-}
-check_functions: dict = {
-
-}
+@configure.utility(provides=IPasswordHasher, name='sha512')
+def sha512_pw_hasher(pw, salt):
+    return hashlib.sha512(pw + salt).hexdigest()
 
 
-def hash_check_password(token, password):
+@configure.utility(provides=IPasswordChecker, name='sha512')
+def hash_password_checker(token, password):
     split = token.split(':')
     if len(split) != 3:
         return False
@@ -41,10 +47,8 @@ def hash_password(password, salt=None, algorithm='sha512'):
     if isinstance(password, str):
         password = password.encode('utf-8')
 
-    hash_func = hash_functions[algorithm]
-    hashed_password = hash_func(password + salt)
-    if hasattr(hashed_password, 'hexdigest'):
-        hashed_password = hashed_password.hexdigest()
+    hash_func = get_utility(IPasswordHasher, name=algorithm)
+    hashed_password = hash_func(password, salt)
     return '{}:{}:{}'.format(algorithm, salt.decode('utf-8'), hashed_password)
 
 
@@ -53,10 +57,10 @@ def check_password(token, password):
     if len(split) != 3:
         return False
     algorithm = split[0]
-    if algorithm in check_functions:
-        check_func = check_functions[algorithm]
-    else:
-        check_func = hash_check_password
+    check_func = query_utility(IPasswordChecker, name=algorithm)
+    if check_func is None:
+        logger.error(f'Could not find password checker for {algorithm}')
+        return False
     return check_func(token, password)